Allen: Started Enemy leveling, implemented HUD and UI assets.
Bobby: "Here is my commit for now I am really in need of the whole map or level to finish the rest of the map.
so for the health of the enemy it is the same as the player it needs the collider or whatever to detect the hit,
so with the assassin if you guys add a item sprite in the level and you give it or them the tag of Item then it will go to steal it, so far I am waiting for the actual level to see the actual items so that I can have the item disappear when the enemy takes it, simple destroy line not hard.
The medic is really complicated more than I thought it takes three arrays one for objects, one for transforms and one for health of every enemy in the area. It takes in all arrays to not only figure out who is the closest enemy but also who has the lowest amount of health and it will attempt to heal them
So the captain and the warrior need the actual level to be implemented the warrior will literally be copy and paste of the assassin with just attacking structures instead of items
The captain does not do much just sort of a bigger harder to kill enemy but I need to have checkpoints for him(need full level).

I really do not have as much left to do as it seems like you guys thought "