Commit- added ObstacleAvoidance.cs so I have been working with this so far it is the closest one to sort of work, all the A* examples I have seen are grid based which is cool if I have to make a grid I will figure it out but I dont think it would look correct for our game.
so what this code does is it makes a simple ray cast in front and behind the enemy.

The ones in front is meant to detect the obstacle from a distance and the one on the back and side is meant to tell the enemy when he is clear of the object.

If this doesnt work about mid next week I will once again try something new also can someone have the actual bases generate again without it I had to make a fake project to test everything.